requirements.txt
APScheduler==3.6.3
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
binance-connector==1.18.0
binance-futures-connector==3.2.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.1
cssselect==1.2.0
cssutils==2.6.0
gevent==22.10.1
greenlet==1.1.3.post0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.11.0
lxml==4.9.1
MarkupSafe==2.1.1
matrix-client==0.4.0
more-itertools==9.0.0
premailer==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
tornado==6.1
twilio==7.15.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
websocket-client==1.2.0
Werkzeug==2.2.2
yagmail==0.15.293
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.0